+++ rocketpy/mathutils/function.py
@@ -1144,6 +1144,51 @@ class Function:
             title=self.title,
         )
 
+    def remove_outliers_iqr(self, threshold=1.5):
+        """Remove outliers from the Function source using the interquartile
+        range method. The Function should have an array-like source.
+
+        Parameters
+        ----------
+        threshold : float, optional
+            Threshold for the interquartile range method. Default is 1.5.
+
+        Returns
+        -------
+        Function
+            The Function with the outliers removed.
+
+        References
+        ----------
+        [1] https://en.wikipedia.org/wiki/Outlier#Tukey's_fences
+        """
+
+        if callable(self.source):
+            raise TypeError(
+                "Cannot remove outliers if the source is a callable object."
+                + " The Function.source should be array-like."
+            )
+
+        x = self.x_array
+        y = self.y_array
+        y_q1 = np.percentile(y, 25)
+        y_q3 = np.percentile(y, 75)
+        y_iqr = y_q3 - y_q1
+        y_lower = y_q1 - threshold * y_iqr
+        y_upper = y_q3 + threshold * y_iqr
+
+        y_filtered = y[(y >= y_lower) & (y <= y_upper)]
+        x_filtered = x[(y >= y_lower) & (y <= y_upper)]
+
+        return Function(
+            source=np.column_stack((x_filtered, y_filtered)),
+            inputs=self.__inputs__,
+            outputs=self.__outputs__,
+            interpolation=self.__interpolation__,
+            extrapolation=self.__extrapolation__,
+            title=self.title,
+        )
+
     # Define all presentation methods
     def __call__(self, *args):
         """Plot the Function if no argument is given. If an
@@ -1474,6 +1519,7 @@ class Function:
         force_data=False,
         force_points=False,
         return_object=False,
+        show=True,
     ):
         """Plots N 1-Dimensional Functions in the same plot, from a lower
         limit to an upper limit, by sampling the Functions several times in
@@ -1481,38 +1527,43 @@ class Function:
 
         Parameters
         ----------
-        plot_list : list
+        plot_list : list[Tuple[Function,str]]
             List of Functions or list of tuples in the format (Function,
             label), where label is a string which will be displayed in the
             legend.
-        lower : scalar, optional
-            The lower limit of the interval in which the Functions are to be
-            plotted. The default value for function type Functions is 0. By
-            contrast, if the Functions given are defined by a dataset, the
-            default value is the lowest value of the datasets.
-        upper : scalar, optional
-            The upper limit of the interval in which the Functions are to be
-            plotted. The default value for function type Functions is 10. By
-            contrast, if the Functions given are defined by a dataset, the
-            default value is the highest value of the datasets.
+        lower : float, optional
+            This represents the lower limit of the interval for plotting the
+            Functions. If the Functions are defined by a dataset, the smallest
+            value from the dataset is used. If no value is provided (None), and
+            the Functions are of Function type, 0 is used as the default.
+        upper : float, optional
+            This represents the upper limit of the interval for plotting the
+            Functions. If the Functions are defined by a dataset, the largest
+            value from the dataset is used. If no value is provided (None), and
+            the Functions are of Function type, 10 is used as the default.
         samples : int, optional
             The number of samples in which the functions will be evaluated for
             plotting it, which draws lines between each evaluated point.
             The default value is 1000.
-        title : string, optional
+        title : str, optional
             Title of the plot. Default value is an empty string.
-        xlabel : string, optional
+        xlabel : str, optional
             X-axis label. Default value is an empty string.
-        ylabel : string, optional
+        ylabel : str, optional
             Y-axis label. Default value is an empty string.
-        force_data : Boolean, optional
+        force_data : bool, optional
             If Function is given by an interpolated dataset, setting force_data
             to True will plot all points, as a scatter, in the dataset.
             Default value is False.
-        force_points : Boolean, optional
+        force_points : bool, optional
             Setting force_points to True will plot all points, as a scatter, in
             which the Function was evaluated to plot it. Default value is
             False.
+        return_object : bool, optional
+            If True, returns the figure and axis objects. Default value is
+            False.
+        show : bool, optional
+            If True, shows the plot. Default value is True.
 
         Returns
         -------
@@ -1586,7 +1637,8 @@ class Function:
         plt.xlabel(xlabel)
         plt.ylabel(ylabel)
 
-        plt.show()
+        if show:
+            plt.show()
 
         if return_object:
             return fig, ax